+++ src/main/java/blusunrize/immersiveengineering/common/blocks/metal/SawmillTileEntity.java
@@ -42,6 +42,7 @@ import net.minecraft.util.NonNullList;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.math.AxisAlignedBB;
 import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.math.MathHelper;
 import net.minecraft.util.math.shapes.ISelectionContext;
 import net.minecraft.util.math.shapes.VoxelShape;
 import net.minecraft.util.math.shapes.VoxelShapes;
@@ -173,6 +174,7 @@ public class SawmillTileEntity extends PoweredMultiblockTileEntity<SawmillTileEn
 					this.sawblade = ItemStack.EMPTY;
 					this.updateMasterBlock(null, true);
 				}
+				this.updateComparatorLevel(false);
 			}
 		}
 		for(ItemStack output : secondaries)
@@ -192,6 +194,7 @@ public class SawmillTileEntity extends PoweredMultiblockTileEntity<SawmillTileEn
 				else if(!world.isRemote)
 					player.entityDropItem(master.sawblade.copy(), 0);
 				master.sawblade = ItemStack.EMPTY;
+				master.updateComparatorLevel(false);
 				this.updateMasterBlock(null, true);
 				return true;
 			}
@@ -211,6 +214,7 @@ public class SawmillTileEntity extends PoweredMultiblockTileEntity<SawmillTileEn
 					else if(!world.isRemote)
 						player.entityDropItem(tempBlade, 0);
 				}
+				master.updateComparatorLevel(false);
 				this.updateMasterBlock(null, true);
 				return true;
 			}
@@ -293,6 +297,47 @@ public class SawmillTileEntity extends PoweredMultiblockTileEntity<SawmillTileEn
 		return ImmutableSet.of(
 				new BlockPos(0, 1, 2)
 		);
+	}
+
+	private int cachedComparatorLevel = -1;
+
+	private void updateComparatorLevel(boolean force)
+	{
+		float damage = 1-(this.sawblade.getDamage()/(float)this.sawblade.getMaxDamage());
+		int level = MathHelper.floor(damage*15);
+
+		if(level!=cachedComparatorLevel||force)
+		{
+			this.cachedComparatorLevel = level;
+			this.markDirty();
+
+			Set<BlockPos> rsPositions = getRedstonePos();
+			for(BlockPos rsPos : rsPositions)
+			{
+				SawmillTileEntity tile = this.getTileForPos(rsPos);
+				if(tile!=null)
+				{
+					tile.cachedComparatorLevel = level;
+					tile.markDirty();
+					tile.markContainingBlockForUpdate(null);
+				}
+			}
+		}
+	}
+
+
+	@Override
+	public int getComparatorInputOverride()
+	{
+		if(!this.isRedstonePos())
+			return 0;
+		if(this.cachedComparatorLevel < 0)
+		{
+			SawmillTileEntity master = master();
+			if(master!=null)
+				master.updateComparatorLevel(true);
+		}
+		return this.cachedComparatorLevel;
 	}
 
 	@Override